# Combine index load with index opening