## No big impact:

```
benchmark                   old ns/op      new ns/op      delta
BenchmarkRebuildIndex-4     3058643446     2980298658     -2.56%
BenchmarkIndexLoad-4        80980738       81759976       +0.96%

benchmark                   old allocs     new allocs     delta
BenchmarkRebuildIndex-4     245708         245600         -0.04%
BenchmarkIndexLoad-4        10976          11005          +0.26%

benchmark                   old bytes     new bytes     delta
BenchmarkRebuildIndex-4     48646032      48623864      -0.05%
BenchmarkIndexLoad-4        46753343      46764222      +0.02%
```